+++ workbench/sveltekit/src/routes/+page.svelte
@@ -1,3 +1,3 @@
-<h1>Workflows DevKit + SvelteKit Example</h1>
+<h1>Workflow SDK + SvelteKit Example</h1>
 <hr />
 <textarea id="output" readonly placeholder="output"></textarea>