+++ packages/nest/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @workflow/nest
 
+## 0.0.0-beta.25
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @workflow/builders@4.0.1-beta.67
+
+## 0.0.0-beta.24
+
+### Patch Changes
+
+- Updated dependencies [[`6dc1b78`](https://github.com/vercel/workflow/commit/6dc1b785822af5c1dc3b4a2a9b1dcb7f626cf5ff)]:
+  - @workflow/builders@4.0.1-beta.66
+
 ## 0.0.0-beta.23
 
 ### Patch Changes